Termino Repository, service y controller de todas las entidades.

+++ APP/BACKEND/AplicacionGuardiasProfesores/src/main/java/com/daw/repositories/HorarioProfesorRepository.java
@@ -6,6 +6,6 @@ import com.daw.datamodel.entities.HorariosProfesor;
 
 
 
-public interface HorarioProfesorRepository extends JpaRepository<HorariosProfesor, Long>{
+public interface HorarioProfesorRepository extends JpaRepository<HorariosProfesor, Integer>{
 
 }
